Create validity bitmap buffer directly from binary
Previously validity bitmap's buffer was create with buffer:new/2 which
expected a list as input. Thus each byte was an element in a list, to
which each element was appended after processing in a loop. This
approach had performance issues as on each iteration, the list was
copied.

With this commit, we use a binary to store each byte, and adapt our
buffer implementation to directly accept binaries. This results in much
better performance.