Move bitmap implementation to its own module
Previously, the only thing in the `serde_arrow_array_utils` was the
validity bitmap implementation. Thus, I decided to rename this module to
`serde_arrow_bitmap` and also write tests for the implementation.

This allows us to have good tests for the bitmap implementation and have
a more "util-like" utils module which is not part of the core
functionality of arrow.