Redo polymorphism by using a single record for all arrays
Previously, all arrays or layouts had a single record for that layout.
The modules of all these arrays would then respect the behaviour of
`serde_arrow_array`. Since Erlang doesn't support Protocols like Elixir
did, this setup was clunky even though protocols are built on top of
regular Erlang behaviours.

Instead, in this commit, I have set things so that all layouts use the
same record definition. The layout is then specified in a field. This is
a much simpler way of solving this problem and allows for much fewer
includes. Any field that does not apply to a layout is marked as
undefined, for example offsets for the primitive layout.