Remove type record inclusion from type test suite
This serde_arrow_type.hrl file is a remainder of some experiments with
expressing types as a record. For some reason, deleting the file on my
laptop and then running the tests didn't fail tests. However, this test
failed on pushing changes in the CI. This commit remove the inclusion.