Store binary data as a single binary in a buffer
Since each element in a list of binaries does not have a definite
length, a null value does not have a length. Thus, we depend on offsets
and validity bitmaps to demarcate a null value.

Additionally from a buffer point of view, when we read from arrow, we
cannot split the binary in a list of binaries without the offsets. Thus
we will have to deal with 2 possibilities on to_erlang:

1. from_erlang -> to_erlang: where the binaries are split properly with
nulls in between

2. from_arrow -> to_arrow: where we have a single element list with a
single binary with space that may or may not be allocated for nulls

Thus we now store list of binaries as a single binary concatenated, so
that we have a unified approach with working with binary data.