Redesign buffer data storage (#16)
This PR redesigns how buffers are used and managed. Firstly, 3 new functions have been introduced:

1. from_erlang/2
2. to_arrow/2
3. to_erlang/1

These will be called when we need to create a buffer from some source, or convert them in to another format.

Secondly, internally in data field, data will be stored in the erlang form and not in the arrow form. This will allow data in buffers to be used later to do something. Additionally, from_erlang -> to_erlang becomes much more efficient as there is not serde involved.

You can have a look at the log for more details:

* Add from_erlang/2 to serde_arrow_buffer

* Add to_arrow/2

* Add to_arrow docs

* Add to_erlang/1

* Store binary data as a single binary in a buffer

Since each element in a list of binaries does not have a definite
length, a null value does not have a length. Thus, we depend on offsets
and validity bitmaps to demarcate a null value.

Additionally from a buffer point of view, when we read from arrow, we
cannot split the binary in a list of binaries without the offsets. Thus
we will have to deal with 2 possibilities on to_erlang:

1. from_erlang -> to_erlang: where the binaries are split properly with
nulls in between

2. from_arrow -> to_arrow: where we have a single element list with a
single binary with space that may or may not be allocated for nulls

Thus we now store list of binaries as a single binary concatenated, so
that we have a unified approach with working with binary data.

* Remove serde_arrow_buffer:from_binary/3

This commit removes the serde_arrow_buffer:from_binary/3 function and
replaces all instances of it with either from_erlang/2 or new/2 where
applicable.

* Remove serde_arrow_buffer:new/2

This commit removes the new/2 function, and replaces all instances of it
with from_erlang/2. It also makes the necessary changes in the tests.

Please do note that the invalid data tests for fixed_lists fail and have
been temporarily commented out.

* Crash on invalid data in fixed_lists

This commit adds a temporary solution for crashing on invalid data in
fixed lists. However, 2 tests, specifically when there is not nesting,
should crash at the primitive array level, and not at the fixed list
level.

* rebar3 fmt

* Fix dialyzer errors

This commit fixes the dialyzer errors. The error stems from the
inability to express nesting of list of native values properly.
Currently I have made things fall back to list() as a temporary
solution, but I will need to define types that deal with this properly.