Add incomplete implementation of from_arrow
This commits adds tests and an incomplete of from_arrow. Right now, it
can manage assigning a type, and length to the buffer, but is unable to
deserialize the actual data from arrow. This is because we also need a
validity bitmap to deserialize data, something which a buffer doesn't
have.