Rename array:new to array:from_erlang
This commit renames array:new, and its callbacks, and its compliant
modules' functions to from_erlang. This is necessary to distinguish
between creating an array from erlang and from arrow. This is also sets
the precedent for conversions to erlang and to arrow. This change is
similar to the change done to buffer.