Serialize Schema Fields
In addition to serializing schema fields, there have been some soft
changes. In creating this commit, I've come to realise that Atoms are a
bad thing to keep in Rust structs. So, in this commit, I've also changed
Name and Dictionary to use enums to represent undefined instead of an
atom.